+++ lib/store_locator.dart
@@ -1,4 +1,3 @@
-library store_locator;
-
+export 'typedef.dart';
 export 'src/map_configuration.dart';
 export 'src/store_locator.dart';